create dataAccesor class and implem
